tests/wpt/tests/css/css-env/at-supports.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <link rel="help" href="https://drafts.csswg.org/css-env-1/">
    <title>Test that CSS env vars work with @support</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      @supports (background-color: env(test)) {
        body { background-color: rgb(0, 128, 0); }
      }
    </style>
  </head>
  <body>
    <script>
    test(() => {
      const style = window.getComputedStyle(document.body);
      assert_equals(style.getPropertyValue("background-color"), "rgb(0, 128, 0)");
    });
    </script>
  </body>
</html>
